refact: Separate SSHandler into SSHConnection and SSHExecution
